+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added `From` implementation for `asset::Symbol` type
 - Converted `Account::daytrade_count` to `u64`
 - Decreased tracing verbosity by one level
 - Bumped `num-decimal` dependency to `0.2`